feat: 200/300번대 공통 로직 Controller 클래스로 추상화, 각 컨트롤러가 extend
