+++ src/main/java/controller/FileController.java
@@ -13,7 +13,7 @@ import static webserver.http.HttpUtil.*;
 import static service.FileService.readStaticFile;
 import static service.SessionService.getUserIdBySid;
 
-public abstract class FileController {
+public abstract class FileController extends Controller {
     protected static final String NAVBAR_RIGHT = "<li>.*?/user/list\\.html.*?</li>";
     protected static final String USERNAME_FORMAT = "<li>%s</li>";
     protected static final String BUTTON_LOGOUT = "<li>.*?로그아웃.*?</li>";
@@ -61,19 +61,5 @@ public abstract class FileController {
         }
 
         return source;
-    }
-
-    protected static Response generate200Response(MIME mime, byte[] body) {
-        Map<String, String> headerMap = new HashMap<>();
-        headerMap.put(HEADER_CONTENT_TYPE, mime.getMime() + HEADER_CHARSET);
-        headerMap.put(HEADER_CONTENT_LENGTH, String.valueOf(body.length));
-
-        return new Response(STATUS.OK, headerMap, body);
-    }
-
-    protected static Response generate303Response(String redirectUrl) {
-        Map<String, String> headerMap = new HashMap<>();
-        headerMap.put(HEADER_REDIRECT_LOCATION, redirectUrl);
-        return new Response(STATUS.SEE_OTHER, headerMap, null);
     }
 }
